Fix some circular imports, ready for py3.
