+++ ckan/model/core.py
@@ -1,28 +1,8 @@
 # encoding: utf-8
 
-from ckan.model import domain_object
-
-
-__all__ = ['System', 'State', 'StatefulObjectMixin']
 log = __import__('logging').getLogger(__name__)
 
 
-class System(domain_object.DomainObject):
-
-    name = 'system'
-
-    def __unicode__(self):
-        return u'<%s>' % self.__class__.__name__
-
-    def purge(self):
-        pass
-
-    @classmethod
-    def by_name(cls, name):
-        return System()
-
-
-# VDM-specific domain objects
 class State(object):
     ACTIVE = u'active'
     DELETED = u'deleted'
