+++ ckan/plugins/__init__.py
@@ -3,4 +3,4 @@
 from ckan.plugins.core import *
 from ckan.plugins.interfaces import *
 
-import toolkit
+from ckan.plugins import toolkit
